платежная система

+++ database/migrations/2024_05_13_100140_create_unreg_orders_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('user_address_info')->nullable();
             $table->text('comment')->nullable();
             $table->text('products');
-            $table->integer('sum');
+            $table->float('sum');
             $table->timestamps();
         });
     }
